requirements.txt
python-dotenv~=1.0.1
typing_extensions~=4.9.0
pillow~=10.2.0
pip~=24.1.2
filelock~=3.13.1
requests~=2.32.3
colorama~=0.4.6
numpy~=1.26.3
llvmlite~=0.43.0
numba~=0.60.0
setuptools~=65.5.1
Jinja2~=3.1.3
regex~=2024.5.15
sympy~=1.12
mpmath~=1.3.0
torch~=2.3.1+cu121
tqdm~=4.66.4
torchvision~=0.18.1+cu121
urllib3~=2.2.2
fsspec~=2024.2.0
MarkupSafe~=2.1.5
certifi~=2024.6.2
tiktoken~=0.7.0
greenlet~=3.0.3
networkx~=3.2.1
idna~=3.7
SQLAlchemy~=2.0.31
pyodbc~=5.1.0
torchaudio~=2.3.1+cu121